Create separate logginginterceptor module
