+++ logginginterceptor/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest package="net.grandcentrix.thirtyinch.logginginterceptor">
+
+    <application />
+
+</manifest>
