+++ settings.gradle
@@ -1,1 +1,2 @@
-include ':thirtyinch', ':plugin', ':sample', ':rx', ':test', ':plugin-test', ':rx2'
+include ':thirtyinch', ':plugin', ':sample', ':rx', ':test', ':plugin-test', ':rx2',
+        ':logginginterceptor'
